requirements.txt
# Framework Web
python-fasthtml>=0.12.12  # Use a versão original ou mais recente
uvicorn[standard]>=0.22.0 # Servidor ASGI

# Dependências dos Módulos Originais (mantenha todas)
aider-install==0.1.3
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==44.0.2
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
docx2pdf==0.1.8
gitdb==4.0.11
GitPython==3.1.42
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
img2pdf==0.6.0
importlib-metadata==7.0.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.4
ocrmypdf==16.10.0
openai==1.13.3
packaging==23.2
pandas==2.2.1
pdfminer.six==20250327
pi_heif==0.22.0
pikepdf==9.5.2
pillow==10.2.0
pluggy==1.5.0
protobuf==4.25.3
pyarrow==15.0.0
pycparser==2.22
pydantic>=1.10,<2.0 # FastHTML pode ter dependência específica, verificar docs se der erro
pydantic_core
pydeck==0.8.1b0
Pygments==2.17.2
PyMuPDF==1.25.5
PyPDF2==3.0.1
python-dateutil>=2.8.0 # Adicionada para calculadora
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
referencing==0.33.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
# streamlit removido
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typing_extensions
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
uv==0.6.12
validators==0.22.0
watchdog==4.0.0
wrapt==1.17.2
zipp==3.17.0
openai-whisper
langchain>=0.1.0
langchain-community>=0.0.1
langchain-openai>=0.1.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
pypdf>=3.0.0
accelerate>=0.20.0